+++ packages/editor-preset-general/CHANGELOG.md
@@ -3,48 +3,88 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< HEAD
 <a name="1.0.0"></a>
 # [1.0.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general-beta@0.14.0...@ali/lowcode-editor-preset-general-beta@1.0.0) (2020-08-17)
+=======
+<a name="0.9.28"></a>
+## [0.9.28](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.27...@ali/lowcode-editor-preset-general@0.9.28) (2020-08-14)
+>>>>>>> master
 
 
 
 
+<<<<<<< HEAD
 **Note:** Version bump only for package @ali/lowcode-editor-preset-general-beta
 
 <a name="0.14.0"></a>
 # [0.14.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general-beta@0.13.0...@ali/lowcode-editor-preset-general-beta@0.14.0) (2020-08-17)
+=======
+**Note:** Version bump only for package @ali/lowcode-editor-preset-general
+
+<a name="0.9.27"></a>
+## [0.9.27](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.26...@ali/lowcode-editor-preset-general@0.9.27) (2020-08-14)
+>>>>>>> master
 
 
 
 
+<<<<<<< HEAD
 **Note:** Version bump only for package @ali/lowcode-editor-preset-general-beta
 
 <a name="0.13.0"></a>
 # [0.13.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general-beta@0.11.0...@ali/lowcode-editor-preset-general-beta@0.13.0) (2020-08-17)
+=======
+**Note:** Version bump only for package @ali/lowcode-editor-preset-general
+
+<a name="0.9.26"></a>
+## [0.9.26](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.25...@ali/lowcode-editor-preset-general@0.9.26) (2020-08-07)
+>>>>>>> master
 
 
 
 
+<<<<<<< HEAD
 **Note:** Version bump only for package @ali/lowcode-editor-preset-general-beta
 
 <a name="0.12.0"></a>
 # [0.12.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general-beta@0.11.0...@ali/lowcode-editor-preset-general-beta@0.12.0) (2020-08-17)
+=======
+**Note:** Version bump only for package @ali/lowcode-editor-preset-general
+
+<a name="0.9.25"></a>
+## [0.9.25](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.24...@ali/lowcode-editor-preset-general@0.9.25) (2020-08-06)
+>>>>>>> master
 
 
 
 
+<<<<<<< HEAD
 **Note:** Version bump only for package @ali/lowcode-editor-preset-general-beta
 
 <a name="0.11.0"></a>
 # 0.11.0 (2020-08-16)
+=======
+**Note:** Version bump only for package @ali/lowcode-editor-preset-general
+
+<a name="0.9.24"></a>
+## [0.9.24](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.23...@ali/lowcode-editor-preset-general@0.9.24) (2020-08-06)
+>>>>>>> master
 
 
 
 
+<<<<<<< HEAD
 **Note:** Version bump only for package @ali/lowcode-editor-preset-general-beta
 
 <a name="0.10.0"></a>
 # [0.10.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.22...@ali/lowcode-editor-preset-general@0.10.0) (2020-08-14)
+=======
+**Note:** Version bump only for package @ali/lowcode-editor-preset-general
+
+<a name="0.9.23"></a>
+## [0.9.23](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-editor-preset-general@0.9.22...@ali/lowcode-editor-preset-general@0.9.23) (2020-08-06)
+>>>>>>> master