+++ packages/rax-simulator-renderer/CHANGELOG.md
@@ -3,14 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< HEAD
 <a name="1.0.0"></a>
 # [1.0.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.15.0...@ali/lowcode-rax-simulator-renderer@1.0.0) (2020-08-17)
+=======
+<a name="0.8.47"></a>
+## [0.8.47](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.46...@ali/lowcode-rax-simulator-renderer@0.8.47) (2020-08-14)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
 
+<<<<<<< HEAD
 <a name="0.15.0"></a>
 # [0.15.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.14.0...@ali/lowcode-rax-simulator-renderer@0.15.0) (2020-08-17)
 
@@ -29,36 +35,67 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 <a name="0.13.0"></a>
 # [0.13.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.12.0...@ali/lowcode-rax-simulator-renderer@0.13.0) (2020-08-17)
+=======
+<a name="0.8.46"></a>
+## [0.8.46](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.45...@ali/lowcode-rax-simulator-renderer@0.8.46) (2020-08-14)
+
+
+### Bug Fixes
+
+* update package.json ([f1ec59c](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/commit/f1ec59c))
+
+
+
+
+<a name="0.8.45"></a>
+## [0.8.45](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.44...@ali/lowcode-rax-simulator-renderer@0.8.45) (2020-08-14)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
 
+<<<<<<< HEAD
 <a name="0.12.0"></a>
 # [0.12.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.11.0...@ali/lowcode-rax-simulator-renderer@0.12.0) (2020-08-16)
+=======
+<a name="0.8.44"></a>
+## [0.8.44](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.43...@ali/lowcode-rax-simulator-renderer@0.8.44) (2020-08-07)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
 
+<<<<<<< HEAD
 <a name="0.11.0"></a>
 # [0.11.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.10.0...@ali/lowcode-rax-simulator-renderer@0.11.0) (2020-08-14)
+=======
+<a name="0.8.43"></a>
+## [0.8.43](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.42...@ali/lowcode-rax-simulator-renderer@0.8.43) (2020-08-06)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
 
+<<<<<<< HEAD
 <a name="0.10.0"></a>
 # [0.10.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.9.0...@ali/lowcode-rax-simulator-renderer@0.10.0) (2020-08-14)
+=======
+<a name="0.8.42"></a>
+## [0.8.42](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.41...@ali/lowcode-rax-simulator-renderer@0.8.42) (2020-08-06)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
 
+<<<<<<< HEAD
 <a name="0.9.0"></a>
 # [0.9.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.40...@ali/lowcode-rax-simulator-renderer@0.9.0) (2020-08-14)
 
@@ -66,6 +103,18 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 **Note:** Version bump only for package @ali/lowcode-rax-simulator-renderer
+=======
+<a name="0.8.41"></a>
+## [0.8.41](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.40...@ali/lowcode-rax-simulator-renderer@0.8.41) (2020-08-06)
+
+
+### Bug Fixes
+
+* 修复 preset-vision 版本 lifeCycles 丢失以及 slot 初始化问题 ([7cf6d24](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/commit/7cf6d24))
+
+
+
+>>>>>>> master
 
 <a name="0.8.40"></a>
 ## [0.8.40](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-rax-simulator-renderer@0.8.39...@ali/lowcode-rax-simulator-renderer@0.8.40) (2020-08-04)